+++ packages/iiif/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zazuko/trifid-plugin-iiif
 
+## 1.0.1
+
+### Patch Changes
+
+- 2bfe266: Remove `@rdfjs/express-handler` dependency
+
 ## 1.0.0
 
 ### Major Changes